--- bullet.py
@@ -1,38 +1,0 @@
-import pygame
-from pygame.sprite import Sprite
-
-
-class Bullet(Sprite):
-    def __init__(self, ai_game):
-        super().__init__()
-        self.screen = ai_game.screen
-        self.settings = ai_game.settings
-        self.color = self.settings.bullet_color
-        self.rect = pygame.Rect(0, 0, self.settings.bullet_width, self.settings.bullet_height)
-        self.rect.midtop = ai_game.ship.rect.midtop
-        self.y = float(self.rect.y)
-
-    def update(self):
-        self.y -= self.settings.bullet_speed
-        self.rect.y = self.y
-
-    def draw_bullet(self):
-        pygame.draw.rect(self.screen, self.color, self.rect)
-
-
-class Rocket(Sprite):
-    def __init__(self, ai_game):
-        super().__init__()
-        self.screen = ai_game.screen
-        self.settings = ai_game.settings
-        self.color = self.settings.rocket_color
-        self.rect = pygame.Rect(0, 0, self.settings.rocket_width, self.settings.rocket_height)
-        self.rect.midtop = ai_game.ship.rect.midtop
-        self.y = float(self.rect.y)
-
-    def update(self):
-        self.y -= self.settings.rocket_speed
-        self.rect.y = self.y
-
-    def draw_rocket(self):
-        pygame.draw.rect(self.screen, self.color, self.rect)
