--- game_stats.py
@@ -1,12 +1,0 @@
-class GameStats():
-    def __init__(self, ai_game):
-        self.settings = ai_game.settings
-        self.reset_stats()
-        self.game_active = False
-        self.high_score = 0
-
-    def reset_stats(self):
-        self.ships_left = self.settings.ship_limit
-        self.rockets_left = self.settings.rocket_limit
-        self.score = 0
-        self.level = 1
