--- sound_effects.py
@@ -1,10 +1,0 @@
-import pygame
-
-
-pygame.mixer.init()
-
-bullet_sound = pygame.mixer.Sound('sounds/bullet.wav')
-rocket_sound = pygame.mixer.Sound('sounds/rocket.wav')
-alien_sound = pygame.mixer.Sound('sounds/explosion.wav')
-ship_collides_sound = pygame.mixer.Sound('sounds/explosion.wav')
-soundtrack_sound = pygame.mixer.Sound('sounds/through space.ogg')
